Rename files and rename TomographicProjector to XRayTransform
